+++ src/install.js
@@ -1,7 +1,0 @@
-import FormLoader from "./components/forms/FormLoader.vue";
-
-export default {
-  install(Vue) {
-    Vue.component("form-loader", FormLoader);
-  }
-};