+++ vue.config.js
@@ -1,3 +1,13 @@
+const { VuetifyLoaderPlugin } = require("vuetify-loader");
+
 module.exports = {
-  chainWebpack: config => config.resolve.set("symlinks", false)
+  chainWebpack: config => config.resolve.set("symlinks", false),
+  configureWebpack: {
+    entry: "./src/index.js",
+    output: {
+      libraryExport: "default"
+    },
+    externals: ["vuetify", "vue", "axios"],
+    plugins: [new VuetifyLoaderPlugin()]
+  }
 };